+++ packages/forgae-cli/forgae-contracts/forgae-contracts.js
