+++ packages/forgae-cli/forgae-init/artifacts/ExampleContract.aes
