+++ packages/forgae-cli/forgae-init/artifacts/contractsAeppSettings.js
