+++ packages/forgae-cli/forgae-init/artifacts/deployTemplate.js
