+++ packages/forgae-cli/forgae-init/artifacts/docker-compose.yml
