+++ packages/forgae-cli/forgae-init/artifacts/docker/aeternity_node1_mean15.yaml
