+++ packages/forgae-cli/forgae-init/artifacts/docker/aeternity_node2_mean15.yaml
