+++ packages/forgae-cli/forgae-init/artifacts/docker/aeternity_node3_mean15.yaml
