+++ packages/forgae-cli/forgae-init/artifacts/docker/entrypoint.sh
