+++ packages/forgae-cli/forgae-init/artifacts/docker/healthcheck.sh
