+++ packages/forgae-cli/forgae-init/artifacts/docker/keys/beneficiary/key.pub
