+++ packages/forgae-cli/forgae-init/artifacts/docker/keys/node1/peer_key
