+++ packages/forgae-cli/forgae-init/artifacts/docker/keys/node1/sign_key
