+++ packages/forgae-cli/forgae-init/artifacts/docker/keys/node2/peer_key
