+++ packages/forgae-cli/forgae-init/artifacts/docker/keys/node2/sign_key.pub
