+++ packages/forgae-cli/forgae-init/artifacts/docker/keys/node3/peer_key.pub
