+++ packages/forgae-cli/forgae-init/artifacts/docker/keys/node3/sign_key
