+++ packages/forgae-cli/forgae-init/artifacts/docker/nginx-default.conf
