+++ packages/forgae-cli/forgae-init/artifacts/docker/nginx-ws.conf
