+++ packages/forgae-cli/forgae-init/artifacts/exampleTests.js
