+++ packages/forgae-cli/forgae-init/artifacts/package.json
