+++ packages/forgae-cli/forgae-init/artifacts/shape/shapeDeployTemplate.js
