+++ packages/forgae-cli/forgae-init/constants.json
