+++ packages/forgae-cli/forgae-shapes/shape-vue-cmd.js
@@ -1,6 +1,6 @@
 const util = require('util');
 const childProcess = util.promisify(require('child_process').exec);
-const utils = require('./../utils');
+const utils = require('../utils');
 const exec = utils.execute;
 
 const fs = require('fs-extra');
