+++ packages/forgae-cli/forgae-test/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "forgae-test",
+  "version": "1.0.0",
+  "description": "",
+  "main": "forgae-test.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC"
+}
