+++ packages/forgae-logger/forgae-history/log-json-store.js
