+++ packages/forgae-logger/forgae-history/log-store-service.js
